requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
DateTime==4.3
html5lib==1.1
icalendar==4.0.7
idna==2.10
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
python-dotenv==0.19.0
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
webencodings==0.5.1
wincertstore==0.2
zope.interface==5.4.0
